Apply PSR-2 and PSR-4 to all classes
* Put each class in its own file per PSR-2

* Autoload all classes using PSR-4
