+++ phpunit.xml
@@ -3,7 +3,7 @@
         bootstrap="./vendor/autoload.php">
     <testsuites>
         <testsuite name="Test Suite">
-            <directory suffix="tests.php">./tests/</directory>
+            <directory suffix=".php">./tests/</directory>
         </testsuite>
     </testsuites>
 </phpunit>
