+++ tests/Fixtures/GripPubControlTestClass.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace GripControl\Test\Fixtures;
+
+use GripControl;
+
+class GripPubControlTestClass extends GripControl\GripPubControl
+{
+    public function getClients()
+    {
+        return $this->clients;
+    }
+
+    public function getPcccbHandlers()
+    {
+        return $this->pcccbhandlers;
+    }
+}
