+++ tests/HttpStreamFormatTests.php
@@ -1,6 +1,10 @@
 <?php
 
-class HttpStreamFormatTests extends PHPUnit_Framework_TestCase
+namespace GripControl\Test;
+
+use GripControl;
+
+class HttpStreamFormatTests extends \PHPUnit_Framework_TestCase
 {
     public function testIntialize()
     {
