+++ tests/ResponseTests.php
@@ -1,6 +1,10 @@
 <?php
 
-class ResponseTests extends PHPUnit_Framework_TestCase
+namespace GripControl\Test;
+
+use GripControl;
+
+class ResponseTests extends \PHPUnit_Framework_TestCase
 {
     public function testIntialize()
     {
