+++ tests/WebSocketEventTests.php
@@ -1,6 +1,10 @@
 <?php
 
-class WebSocketEventTests extends PHPUnit_Framework_TestCase
+namespace GripControl\Test;
+
+use GripControl;
+
+class WebSocketEventTests extends \PHPUnit_Framework_TestCase
 {
     public function testInitialize()
     {
